+++ src/functions/special/Constant.java
@@ -4,6 +4,7 @@ import config.Settings;
 import functions.GeneralFunction;
 import org.jetbrains.annotations.Nullable;
 import tools.ParsingTools;
+import tools.exceptions.IllegalNameException;
 
 import java.util.HashMap;
 import java.util.Map;
@@ -77,6 +78,9 @@ public class Constant extends SpecialFunction {
 	 * @return the value of the constant for convenience
 	 */
 	public static double addSpecialConstant(String string, double value) {
+		if (!Variable.validVariables.matcher(string).matches())
+			throw new IllegalNameException(string);
+		string = ParsingTools.processEscapes(string);
 		specialConstants.put(string, value);
 		return value;
 	}
@@ -132,6 +136,13 @@ public class Constant extends SpecialFunction {
 	public boolean equalsFunction(GeneralFunction that) {
 		return (that instanceof Constant) && (Math.abs(constant - ((Constant) that).constant) < Settings.equalsMargin);
 	}
+
+	public static void resetConstants() {
+		specialConstants.clear();
+		specialConstants.put("π", Math.PI);
+		specialConstants.put("e", Math.E);
+	}
+
 
 	@SuppressWarnings({"VariableNotUsedInsideIf", "ConstantConditions"})
 	public int compareSelf(GeneralFunction that) {