+++ tests/KeywordInterfaceTest.java
@@ -1,8 +1,16 @@
 import config.Settings;
 import functions.GeneralFunction;
+import functions.binary.Pow;
+import functions.commutative.Product;
+import functions.commutative.Sum;
+import functions.special.Constant;
+import functions.special.Variable;
+import functions.unitary.trig.normal.Cos;
+import functions.unitary.trig.normal.Sin;
 import org.junit.jupiter.api.Test;
 import parsing.FunctionParser;
 import parsing.KeywordInterface;
+import tools.DefaultFunctions;
 
 import java.util.List;
 
@@ -114,6 +122,57 @@ public class KeywordInterfaceTest {
     void basicNumericalIntegrationWithError() {
         double[] test = (double[]) KeywordInterface.useKeywords("intne \\sin(x) 0 \\pi");
         assertArrayEquals(new double[]{2, 0.01}, test, 0.01);
+    }
+
+    @Test
+    void basicApostropheParsing() {
+        GeneralFunction test = (GeneralFunction) KeywordInterface.useKeywords("def \\f' 2x");
+        assertEquals(new Product(new Constant(2), new Variable("x")), test);
+    }
+
+    @Test
+    void basicAlphaNumerics() {
+        GeneralFunction test = (GeneralFunction) KeywordInterface.useKeywords("\\ts543s");
+        if (Settings.removeEscapes)
+            assertEquals(new Variable("ts543s"), test);
+        else
+            assertEquals(new Variable("\\ts543s"), test);
+    }
+
+    @Test
+    void basicUnderscoreParsing() {
+        GeneralFunction test = (GeneralFunction) KeywordInterface.useKeywords("\\f2_4_5");
+        if (Settings.removeEscapes)
+            assertEquals(new Variable("f2_4_5"), test);
+        else
+            assertEquals(new Variable("\\f2_4_5"), test);
+    }
+
+    @Test
+    void basicVariableParsing() {
+        GeneralFunction test = (GeneralFunction) KeywordInterface.useKeywords("\\test");
+        if (Settings.removeEscapes)
+            assertEquals(new Variable("test"), test);
+        else
+            assertEquals(new Variable("\\test"), test);
+    }
+
+    @Test
+    void integralEscapedVariable() {
+        GeneralFunction test = (GeneralFunction) KeywordInterface.useKeywords("simp \\int(2\\test)\\d\\test");
+        if (Settings.removeEscapes)
+            assertEquals(DefaultFunctions.square(new Variable("test")), test);
+        else
+            assertEquals(DefaultFunctions.square(new Variable("\\test")), test);
+    }
+
+    @Test
+    void basicTrigExample() {
+        GeneralFunction test = (GeneralFunction) KeywordInterface.useKeywords("def \\trig (\\sin(\\number1 * \\theta))^2 + (\\cos(\\number2 * \\theta))^2 ");
+        if (Settings.removeEscapes)
+            assertEquals(new Sum(new Pow(DefaultFunctions.TWO, new Sin(new Product(new Variable("number1"), new Variable("θ")))), new Pow(DefaultFunctions.TWO, new Cos(new Product(new Variable("number2"), new Variable("θ"))))), test);
+        else
+            assertEquals(new Sum(new Pow(DefaultFunctions.TWO, new Sin(new Product(new Variable("\\number1"), new Variable("θ")))), new Pow(DefaultFunctions.TWO, new Cos(new Product(new Variable("\\number2"), new Variable("θ"))))), test);
     }
 }